Wire up search functionality inside GalleryView [WIP]
This is the my attempt at getting the search functionality to work. It works reasonably well, until the current image is deleted: then the selection moves to the next image of the full set, not of what's currently visible.

Selection state is stored in SDImage, but is updated by ImageStore, while some actions to do it are in ImageController; while only GalleryView knows about the filtered list. This needs to be consolidated.