Implement and improve search functionality (#163)
* Wire up search functionality inside GalleryView

This is the my attempt at getting the search functionality to work. It works reasonably well, until the current image is deleted: then the selection moves to the next image of the full set, not of what's currently visible.

Selection state is stored in SDImage, but is updated by ImageStore, while some actions to do it are in ImageController; while only GalleryView knows about the filtered list. This needs to be consolidated.

* Move selection state from SDImages to ImageStore

* Move filtering to ImageStore

* Remove manual Binding

* Move prev/next navigation to ImageStore to handle filters and wrapping around

* add animation

* quick look

---------

Co-authored-by: Joshua Park <jpark@duck.com>