fix: match tokenizer to current selected model
Since `loadPipeline()` is no longer called whenever a new model is
selected, it is is no longer appropriate to update the tokenizer there.

Instead, the tokenizer is a property of SDModel (which makes sense
because the tokenizer is part of the model data) and is retrieved
directly from the current selected model in ImageController.